Add NovaClientWrapper for using it in NovaCompute
Added a python script for using the Novaclient API directly if the caller is
a Python script, and adding a wrapper for CLI if the script is called by bash.

For example, a bash script can call this way:
 python nova_client.py --os... -os... service list

which will wrap the call to nova service-list by using the novaclient API and
returning a stable output.