Compute: Remove fence_agent optimization
Since we must initiate evacuations from the RA for correctness (because
the control plane might not be healthy when the FA is called), we can't
call the FA at all in case it actually worked.

If that was the case, then there would be two evacuate calls for each
instance and it is a strong possibility that the second would
interrupt/break/abort the first and leave the instance in a broken
state.